requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
setuptools==67.8.0
requests==2.29.0
python-dotenv==1.0.1
pyrebase4==4.7.1
yfinance==0.2.18
gunicorn==21.2.0
